Describe RISC-V register groups
Describe RISC-V register groups based on [1] specification. To avoid a
large amount of groups with few registers each registers are mostly
split into groups based on privilige level, using the first letter in
the name as a hint: `h` for hypervisor, `m` for machine and so on.
Counter/timer and memory protection registers are split into separate
groups because there are enough of them to justify that.

[1] https://github.com/riscv/riscv-isa-manual/releases/download/Priv-v1.12/riscv-privileged-20211203.pdf